devtools/client/shared/components/test/mochitest/test_tabs_accessibility.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test tabs accessibility.
-->
<head>
  <meta charset="utf-8">
  <title>Tabs component accessibility test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = async function () {
  try {
    const ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    const { createFactory } = browserRequire("devtools/client/shared/vendor/react");
    const InspectorTabPanel = createFactory(browserRequire("devtools/client/inspector/components/InspectorTabPanel"));
    const Tabbar =
      createFactory(browserRequire("devtools/client/shared/components/tabs/TabBar"));
    const tabbar = Tabbar();
    const tabbarReact = ReactDOM.render(tabbar, window.document.body);
    const tabbarEl = ReactDOM.findDOMNode(tabbarReact);

    // Setup for InspectorTabPanel
    const tabpanels = document.createElement("div");
    tabpanels.id = "tabpanels";
    document.body.appendChild(tabpanels);

    await addTabWithPanel(0);
    await addTabWithPanel(1);

    const tabAnchors = tabbarEl.querySelectorAll("li.tabs-menu-item a");

    is(tabAnchors[0].parentElement.getAttribute("role"), "presentation", "li role is set correctly");
    is(tabAnchors[0].getAttribute("role"), "tab", "Anchor role is set correctly");
    is(tabAnchors[0].getAttribute("aria-selected"), "true", "Anchor aria-selected is set correctly by default");
    is(tabAnchors[0].getAttribute("aria-controls"), "sidebar-0-panel", "Anchor aria-controls is set correctly");
    is(tabAnchors[1].parentElement.getAttribute("role"), "presentation", "li role is set correctly");
    is(tabAnchors[1].getAttribute("role"), "tab", "Anchor role is set correctly");
    is(tabAnchors[1].getAttribute("aria-selected"), "false", "Anchor aria-selected is set correctly by default");
    is(tabAnchors[1].getAttribute("aria-controls"), "sidebar-1-panel", "Anchor aria-controls is set correctly");

    await setState(tabbarReact, Object.assign({}, tabbarReact.state, {
      activeTab: 1
    }));

    is(tabAnchors[0].getAttribute("aria-selected"), "false", "Anchor aria-selected is reset correctly");
    is(tabAnchors[1].getAttribute("aria-selected"), "true", "Anchor aria-selected is reset correctly");

    function addTabWithPanel(tabId) {
      // Setup for InspectorTabPanel
      let panel = document.createElement("div");
      panel.id = `sidebar-${tabId}`;
      document.body.appendChild(panel);

      return setState(tabbarReact, Object.assign({}, tabbarReact.state, {
        tabs: tabbarReact.state.tabs.concat({
          id: `sidebar-${tabId}`,
          title: `tab-${tabId}`,
          panel: InspectorTabPanel
        }),
      }));
    }
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
};
</script>
</pre>
</body>
</html>
